code/Floatchat/backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.6.0
python-multipart>=0.0.6
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
pandas>=2.2.0
numpy>=1.26.0
netcdf4>=1.6.5
xarray>=2024.1.0
dask>=2024.1.0
scikit-learn>=1.4.0
faiss-cpu>=1.7.4
chromadb>=0.4.18
openai>=1.6.1
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
sentence-transformers>=2.2.2
huggingface-hub>=0.19.4
plotly==5.17.0
folium==0.15.0
requests==2.31.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
websockets==12.0
python-socketio==5.10.0
eventlet==0.33.3
gunicorn==21.2.0
